ReactQuery caching으로 찜하기 기능 로직 수정

+++ src/entities/user/hooks/useWishList.js
@@ -1,4 +1,4 @@
-import { mockWishListData } from '@/pages/single-product/model/mockWishListData';
+import { mockWishlistData } from '@/pages/single-product/model/mockWishListData';
 import { carHarttApi } from '@/shared/api/axios';
 import { useMockToggle } from '@/shared/config/MockToggleProvider';
 import { useMutation, useQuery, useQueryClient } from '@tanstack/react-query';
@@ -6,9 +6,20 @@ import { useMutation, useQuery, useQueryClient } from '@tanstack/react-query';
 // 상품이 찜 목록에 있는지 확인
 // GET /v1/wishes/{itemId}/status
 export function useWishlistStatus(itemId) {
+  const queryClient = useQueryClient();
+  const { useMock } = useMockToggle();
+
   return useQuery({
     queryKey: ['wishlist', 'status', itemId],
     queryFn: async () => {
+      // mock 데이터 사용
+      if (useMock) {
+        // caching된 wishlist 가져오기 (없으면 mock 데이터 fallback)
+        const wishlist =
+          queryClient.getQueryData(['wishlist']) ?? mockWishlistData;
+        const wished = wishlist.some((item) => item.id === itemId);
+        return { wished };
+      }
       const response = await carHarttApi({
         method: 'GET',
         url: `/v1/wishes/${itemId}/status`,
@@ -23,14 +34,21 @@ export function useWishlistStatus(itemId) {
 // 찜 목록 불러오기
 // GET /v1/wishes
 export function useWishlist() {
+  const queryClient = useQueryClient();
   const { useMock } = useMockToggle();
 
   return useQuery({
     queryKey: ['wishlist'],
     queryFn: async () => {
+      // mock 데이터 사용
       if (useMock) {
+        // 최초 mock 데이터 caching
+        const cached = queryClient.getQueryData(['wishlist']);
+        if (cached) return cached;
+
         await new Promise((resolve) => setTimeout(resolve, 1000));
-        return mockWishListData;
+        queryClient.setQueryData(['wishlist'], mockWishlistData);
+        return mockWishlistData;
       }
 
       const response = await carHarttApi({
@@ -47,9 +65,13 @@ export function useWishlist() {
 // POST /v1/wishes
 export function useAddWishlist() {
   const queryClient = useQueryClient();
+  const { useMock } = useMockToggle();
 
   return useMutation({
     mutationFn: async (itemId) => {
+      if (useMock) {
+        return { itemId };
+      }
       const response = await carHarttApi({
         method: 'POST',
         url: '/v1/wishes',
@@ -58,16 +80,28 @@ export function useAddWishlist() {
       });
       return response.data;
     },
-    onSuccess: (data, itemId) => {
+    onMutate: async (itemId) => {
+      await queryClient.cancelQueries(['wishlist']);
+
+      // 롤백용 리스트
+      const prev = queryClient.getQueryData(['wishlist']) || mockWishlistData;
+
+      // ui 업데이트 리스트 (Optimistic Update)
+      const next = [...prev, { id: itemId }];
+      queryClient.setQueryData(['wishlist'], next);
+      // (기대)성공 결과 선반영
       queryClient.setQueryData(['wishlist', 'status', itemId], {
         wished: true,
       });
-      queryClient.invalidateQueries({ queryKey: ['wishlist'] });
+
+      // 서버로부터 실패 응답을 받았을 경우 롤백하기 위해 prev 반환
+      return { prev };
     },
-    onError: (error, itemId) => {
-      queryClient.invalidateQueries({
-        queryKey: ['wishlist', 'status', itemId],
-      });
+    onError: (error, itemId, context) => {
+      if (context?.prev) queryClient.setQueryData(['wishlist'], context.prev);
+    },
+    onSettled: () => {
+      queryClient.invalidateQueries(['wishlist']);
     },
   });
 }
@@ -76,9 +110,14 @@ export function useAddWishlist() {
 // DELETE /v1/wishes/{itemId}
 export function useRemoveWishlist() {
   const queryClient = useQueryClient();
+  const { useMock } = useMockToggle();
 
   return useMutation({
     mutationFn: async (itemId) => {
+      if (useMock) {
+        return { itemId };
+      }
+
       const response = await carHarttApi({
         method: 'DELETE',
         url: `/v1/wishes/${itemId}`,
@@ -86,16 +125,28 @@ export function useRemoveWishlist() {
       });
       return response.data;
     },
-    onSuccess: (data, itemId) => {
+    onMutate: async (itemId) => {
+      await queryClient.cancelQueries(['wishlist']);
+
+      // 롤백용 리스트
+      const prev = queryClient.getQueryData(['wishlist']) || mockWishlistData;
+
+      // ui 업데이트 리스트 (Optimistic Update)
+      const next = prev.filter((item) => item.id === itemId);
+      queryClient.setQueryData(['wishlist'], next);
+      // (기대)성공 결과 선반영
       queryClient.setQueryData(['wishlist', 'status', itemId], {
         wished: false,
       });
-      queryClient.invalidateQueries({ queryKey: ['wishlist'] });
+
+      // 서버로부터 실패 응답을 받았을 경우 롤백하기 위해 prev 반환
+      return { prev };
     },
-    onError: (error, itemId) => {
-      queryClient.invalidateQueries({
-        queryKey: ['wishlist', 'status', itemId],
-      });
+    onError: (error, itemId, context) => {
+      if (context?.prev) queryClient.setQueryData(['wishlist'], context.prev);
+    },
+    onSettled: () => {
+      queryClient.invalidateQueries(['wishlist']);
     },
   });
 }
